backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
supabase
PyJWT
